Update report folder structure
