+++ aleapp.py
@@ -3,14 +3,17 @@ import argparse
 import io
 import os.path
 import typing
-import plugin_loader
 import scripts.report as report
 import traceback
+import sys
+
+import scripts.plugin_loader as plugin_loader
 
 from scripts.search_files import *
 from scripts.ilapfuncs import *
 from scripts.version_info import aleapp_version
 from time import process_time, gmtime, strftime, perf_counter
+from scripts.lavafuncs import *
 
 def validate_args(args):
     if args.artifact_paths or args.create_profile_casedata:
@@ -152,24 +155,30 @@ def main():
     parser.add_argument('-p', '--artifact_paths', required=False, action="store_true",
                         help=("Generate a text file list of artifact paths. "
                               "This argument is meant to be used alone, without any other arguments."))
+    parser.add_argument('--custom_output_folder', required=False, action="store", help="Custom name for the output folder")
 
     loader = plugin_loader.PluginLoader()
     available_plugins = list(loader.plugins)
     profile_filename = None
     casedata = {}
 
+    # Check if no arguments were provided
+    if len(sys.argv) == 1:
+        parser.print_help(sys.stderr)
+        sys.exit()
+
+    args = parser.parse_args()
+
     plugins = []
     plugins_parsed_first = []
 
     for plugin in available_plugins:
-        if plugin.name == 'usagestatsVersion':
+        if plugin.module_name == 'usagestatsVersion':
             plugins_parsed_first.append(plugin)
         else:
             plugins.append(plugin)
 
     selected_plugins = plugins.copy()
-
-    args = parser.parse_args()
 
     try:
         validate_args(args)
@@ -276,6 +285,7 @@ def main():
     wrap_text = args.wrap_text
     output_path = os.path.abspath(args.output_path)
     time_offset = args.timezone
+    custom_output_folder = args.custom_output_folder
 
     # Android file system extractions contain paths > 260 char, which causes problems
     # This fixes the problem by prefixing \\?\ on each windows path.
@@ -283,12 +293,15 @@ def main():
         if input_path[1] == ':' and extracttype =='fs': input_path = '\\\\?\\' + input_path.replace('/', '\\')
         if output_path[1] == ':': output_path = '\\\\?\\' + output_path.replace('/', '\\')
 
-    out_params = OutputParameters(output_path)
+    out_params = OutputParameters(output_path, custom_output_folder)
 
     selected_plugins = plugins_parsed_first + selected_plugins
     
+    initialize_lava(input_path, out_params.report_folder_base, extracttype)
+
     crunch_artifacts(selected_plugins, extracttype, input_path, out_params, wrap_text, loader, casedata, time_offset, profile_filename)
 
+    lava_finalize_output(out_params.report_folder_base)
 
 def crunch_artifacts(
         plugins: typing.Sequence[plugin_loader.PluginSpec], extracttype, input_path, out_params, wrap_text,
@@ -308,13 +321,13 @@ def crunch_artifacts(
     seeker = None
     try:
         if extracttype == 'fs':
-            seeker = FileSeekerDir(input_path)
+            seeker = FileSeekerDir(input_path, out_params.data_folder)
 
         elif extracttype in ('tar', 'gz'):
-            seeker = FileSeekerTar(input_path, out_params.temp_folder)
+            seeker = FileSeekerTar(input_path, out_params.data_folder)
 
         elif extracttype == 'zip':
-            seeker = FileSeekerZip(input_path, out_params.temp_folder)
+            seeker = FileSeekerZip(input_path, out_params.data_folder)
 
         else:
             logfunc('Error on argument -o (input type)')
@@ -343,6 +356,8 @@ def crunch_artifacts(
 
     # Search for the files per the arguments
     for plugin in plugins:
+        logfunc()
+        logfunc('{} [{}] artifact started'.format(plugin.name, plugin.module_name))
         if isinstance(plugin.search, list) or isinstance(plugin.search, tuple):
             search_regexes = plugin.search
         else:
@@ -364,12 +379,10 @@ def crunch_artifacts(
                 log.write(f'</li></ul>')
                 files_found.extend(found)
         if files_found:
-            logfunc()
-            logfunc('{} [{}] artifact started'.format(plugin.name, plugin.module_name))
-            category_folder = os.path.join(out_params.report_folder_base, plugin.category)
+            category_folder = os.path.join(out_params.report_folder_base, '_HTML', plugin.category)
             if not os.path.exists(category_folder):
                 try:
-                    os.mkdir(category_folder)
+                    os.makedirs(category_folder)
                 except (FileExistsError, FileNotFoundError) as ex:
                     logfunc('Error creating {} report directory at path {}'.format(plugin.name, category_folder))
                     logfunc('Error was {}'.format(str(ex)))
@@ -381,11 +394,12 @@ def crunch_artifacts(
                 logfunc('Error was {}'.format(str(ex)))
                 logfunc('Exception Traceback: {}'.format(traceback.format_exc()))
                 continue  # nope
-
-            logfunc('{} [{}] artifact completed'.format(plugin.name, plugin.module_name))
-
+        else:
+            logfunc(f"No file found")
+        logfunc('{} [{}] artifact completed'.format(plugin.name, plugin.module_name))
     log.close()
 
+    write_device_info()
     logfunc('')
     logfunc('Processes completed.')
     end = process_time()
@@ -406,11 +420,11 @@ def crunch_artifacts(
         if input_path.startswith('\\\\?\\'):
             input_path = input_path[4:]
     
-        
-    report.generate_report(out_params.report_folder_base, run_time_secs, run_time_HMS, extracttype, input_path, casedata)
+    report.generate_report(out_params.report_folder_base, run_time_secs, run_time_HMS, extracttype, input_path, casedata, profile_filename, icons)
     logfunc('Report generation Completed.')
     logfunc('')
     logfunc(f'Report location: {out_params.report_folder_base}')
+
     return True
 
 if __name__ == '__main__':
