+++ scripts/modules_to_exclude.py
@@ -1,0 +1,6 @@
+# This list contains the filenames of artifact scripts that take a long time to run.
+# These modules are deselected by default in the GUI.
+
+modules_to_exclude = [
+    'walStrings'
+]
