+++ scripts/report.py
@@ -1,911 +1,13 @@
 import html
 import os
-import pathlib
+from pathlib import Path
 import shutil
 
 from collections import OrderedDict
 from scripts.html_parts import *
 from scripts.ilapfuncs import logfunc
 from scripts.version_info import aleapp_version, aleapp_contributors
-
-# Icon Mappings Dictionary
-# The icon_mappings dictionary is organized by category and is used to map categories and artifacts to icons.
-# Please maintain the list in alphabetical order by category for ease of navigation and updating.
-#
-# To specify an icon for a particular artifact within a category, use the structure:
-# 'CATEGORY': {'ARTIFACT': 'icon-name', ...}
-# Example:
-# 'CHROMIUM': {'BOOKMARKS': 'bookmark', 'DOWNLOADS': 'download', ...}
-#
-# To specify a default icon for all artifacts within a category or a single icon for the entire category, use:
-# 'CATEGORY': 'icon-name' or 'CATEGORY': {'default': 'icon-name', ...}
-# Example:
-# 'ADDRESS BOOK': 'book-open'
-#
-# If a category or artifact does not have a specified icon, the 'alert-triangle' icon will be used as a fallback.
-#
-# Icons are sourced from Feather Icons (feathericons.com). When adding a new icon, ensure that the icon name
-# matches the name listed on the Feather Icons website.
-#
-# The optional '_mode' key can be used to specify a search mode for finding matching artifacts within a category:
-# 'CATEGORY': {'_mode': 'search', ...}
-# In search mode, the function will attempt to find a partial match for the artifact name within the specified category.
-
-icon_mappings = \
-{
-    'ACCESSORY DATA HYUNDAI': 'settings',
-    'ACCOUNT': {
-        'AUTH': 'key',
-        'default': 'user',
-        '_mode': 'search',
-    },
-    'ADB HOSTS': 'terminal',
-    'ADDRESS BOOK': 'book-open',
-    'ADIDAS-RUNNING': {
-        'ACTIVITIES': 'activity',
-        'GOALS': 'target',
-        'USER': 'user',
-        'default': 'user'
-    },
-    'AGGREGATE DICTIONARY': 'book',
-    'AIRDROP DISCOVERABLE': 'search',
-    'AIRDROP EMAILS': 'send',
-    'AIRDROP NUMBERS': 'smartphone',
-    'AIRDROP REAL NAMES': 'user',
-    'AIRTAG DETECTION': 'alert-circle',
-    'AIRTAGS': 'map-pin',
-    'ALARMS': 'clock',
-    'ALFA ROMEO CONTACTS': 'users',  # TODO: adjust artifact to share a category
-    'ALFA ROMEO BLUETOOTH': 'bluetooth',
-    'ALFA ROMEO SIRIUS DATA': 'settings',
-    'ALLTRAILS': {
-        'ALLTRAILS - TRAIL DETAILS': 'map',
-        'ALLTRAILS - USER INFO': 'user',
-    },
-    'ANDROID SYSTEM INTELLIGENCE': {
-        'SIMPLESTORAGE': 'loader',
-        'default': 'user'
-    },
-    'ANDROID NOTIFICATION HISTORY': {
-        'STATUS': 'toggle-left',
-        'NOTIFICATIONS': 'bell',
-        'SNOOZED NOTIFICATIONS': 'bell-off',
-        '_mode': 'search',
-    },
-    'APP CONDUIT': 'activity',
-    'APP INTERACTION': 'bar-chart-2',
-    'APP PERMISSIONS': 'key',
-    'APP ROLES': 'tool',
-    'APP UPDATES': 'codepen',
-    'APPLE MAIL': 'mail',
-    'APPLE NOTES': 'book-open',
-    'APPLE PODCASTS': 'play-circle',
-    'APPLE WALLET': {
-        'CARDS': 'credit-card',
-        'PASSES': 'send',
-        'TRANSACTIONS': 'dollar-sign',
-        'default': 'credit-card',
-    },
-    'APPLICATIONS': 'grid',
-    'AUDIO UUIDS': 'smartphone',
-    'BADOO': {
-        'CHAT': 'message-circle',
-        'CONNECTIONS': 'heart',
-        'default': 'user'
-    },
-    'BASH HISTORY': 'terminal',
-    'BIOME': 'eye',
-    'BIOME APP INSTALL': 'eye',
-    'BIOME BACKLIGHT': 'eye',
-    'BIOME BATTERY PERC': 'eye',
-    'BIOME BLUETOOTH': 'eye',
-    'BIOME CARPLAY CONN': 'eye',
-    'BIOME DEVICE PLUG': 'eye',
-    'BIOME HARDWARE': 'eye',
-    'BIOME IN FOCUS': 'eye',
-    'BIOME INTENTS': 'eye',
-    'BIOME LOCATION ACT': 'eye',
-    'BIOME NOTES': 'eye',
-    'BIOME NOTIFICATIONS PUB': 'eye',
-    'BIOME NOW PLAYING': 'eye',
-    'BIOME SAFARI': 'eye',
-    'BIOME SYNC': 'smartphone',
-    'BIOME TEXT INPUT': 'eye',
-    'BIOME USER ACT META': 'eye',
-    'BIOME WIFI': 'eye',
-    'BLUESKY': {
-        'BLUESKY ACTORS': 'users',
-        'BLUESKY MESSAGES': 'message-circle',
-        'BLUESKY SEARCHES': 'search',
-        'BLUESKY FEED POSTS': 'arrow-down-circle',
-        'BLUESKY POSTS': 'arrow-up-circle',
-        'default': 'cloud',
-    },
-    'BITTORRENT': 'share',
-    'BLUETOOTH': 'bluetooth',
-    'BLUETOOTH CONNECTIONS': 'bluetooth',
-    'BLUETOOTH_DEVICES': 'bluetooth',  # TODO: can this be combined?
-    'BROWSER CACHE': {
-        'CHROME BROWSER CACHE': 'chrome',
-        'default': 'globe'
-    },
-    'BT REPORT': {
-        'GPS DETAIL': 'map-pin',
-        'BT CALL REPORT': 'bluetooth',
-        'default': 'bluetooth',
-    },
-    'BUMBLE': {
-        'BUMBLE - CHAT MESSAGES': 'message-circle',
-        'BUMBLE - USER SETTINGS': 'user',
-        'BUMBLE - MATCHES': 'smile',
-    },
-    'BURNER': {
-        'NUMBER INFORMATION': 'user',
-        'COMMUNICATION INFORMATION': 'message-circle',
-        'default': 'user'
-    },
-    'CACHE DATA': 'box',
-    'CALCULATOR LOCKER': 'lock',
-    'CALENDAR': 'calendar',
-    'CALL HISTORY': {
-        'CALL HISTORY': 'phone-call',
-        'DELETED VOICEMAIL': 'mic-off',
-        'VOICEMAIL': 'mic',
-        'default': 'phone',
-    },
-    'CALL LOGS': 'phone',
-    'CAST': 'cast',
-    'CARPLAY': 'package',
-    'CASH APP': 'credit-card',
-    'CELLULAR WIRELESS': 'bar-chart',
-    'CHASE RETURNS': 'paperclip',
-    'CHATS': 'message-circle',
-    'CHATGPT': {
-       'CONVERSATIONS METADATA': 'info',
-       'CONVERSATIONS': 'message-circle',
-       'USER': 'user',
-       'ACCOUNT USER STATE': 'activity',
-       'ACCOUNT STATUS': 'activity',
-       'CUSTOM INSTRUCTIONS' : 'sliders',
-       'USER SETTINGS' : 'settings',
-       'MEDIA UPLOADS' : 'image',
-       'USER ANALYTICS' : 'activity',
-       '_mode': 'search',
-    },
-    'CHROMIUM': {
-        'AUTOFILL': 'edit-3',
-        'BOOKMARKS': 'bookmark',
-        'DETECT INCIDENTAL PARTY STATE': 'skip-forward',
-        'DOWNLOADS': 'download',
-        'LOGIN': 'log-in',
-        'MEDIA HISTORY': 'video',
-        'NETWORK ACTION PREDICTOR': 'type',
-        'OFFLINE PAGES': 'cloud-off',
-        'SEARCH TERMS': 'search',
-        'TOP SITES': 'list',
-        'WEB VISITS': 'globe',
-        'default': 'chrome',
-        '_mode': 'search',
-    },
-    'CLIPBOARD': 'clipboard',
-    'CLOUDKIT': {
-        'NOTE SHARING': 'share-2',
-        'PARTICIPANTS': 'user',
-    },
-    'COINBASE ARCHIVE': {
-        '3RD': 'log-in',
-        'CARD': 'credit-card',
-        'PERSONAL': 'user',
-        'SITE': 'activity',
-        'TRANS': 'archive',
-        'default': 'monitor',
-    },
-    'CONNECTED DEVICES': 'smartphone',
-    'CONNECTED TO': 'zap',
-    'CONTACTS': 'user',
-    'CONTACT_LIST': 'users',  # TODO: can this use another category?
-    'CONTROL CENTER': {
-        'CONTROL CENTER - ACTIVE CONTROLS': 'sliders',
-        'CONTROL CENTER - DISABLED CONTROLS': 'x-square',
-        'CONTROL CENTER - USER TOGGLED CONTROLS': 'check-square',
-    },
-    'CORE ACCESSORIES': {
-        'ACCESSORYD': 'zap',
-        'USER EVENT AGENT': 'activity',
-    },
-    'COREDUET': {
-        'AIRPLANE MODE': 'pause',
-        'LOCK STATE': 'lock',
-        'PLUGGED IN': 'battery-charging',
-    },
-    'DAHUA TECHNOLOGY (DMSS)': {
-        'CHANNELS': 'film',
-        'DEVICES': 'tablet',
-        'INFO': 'settings',
-        'NOTIFICATIONS': 'bell',
-        'PIN': 'unlock',
-        'SENSORS': 'smartphone',
-        'USER CREATED MEDIA': 'video',
-        '_mode': 'search',
-    },
-    'DATA USAGE': 'wifi',
-    'DEVICE DATA': 'file',
-    'DEVICE HEALTH SERVICES': {
-        'BLUETOOTH': 'bluetooth',
-        'BATTERY': 'battery-charging',
-        'default': 'bar-chart-2',
-        '_mode': 'search'
-    },
-    'DEVICE INFO': {
-        'BUILD INFO': 'terminal',
-        'ANDROID SYSTEM VERSION': 'git-commit',
-        'PARTNER SETTINGS': 'settings',
-        'SETTINGS_SECURE_': 'settings',
-        'default': 'info',
-        '_mode': 'search',
-    },
-    'DHCP': 'settings',
-    'DIAGNOSTIC_DATA': 'thermometer',
-    'DIGITAL WELLBEING': {
-        'ACCOUNT DATA': 'user',
-        'default': 'layers',
-        '_mode': 'search',
-    },
-    'DIGITAL WELLBEING ACCOUNT': {
-        'ACCOUNT DATA': 'user',
-        'default': 'layers',
-        '_mode': 'search',
-    },
-    'DISCORD': {
-        'DISCORD ACCOUNT': 'user',
-        'DISCORD MANIFEST': 'file-text',
-        'DISCORD MESSAGES': 'message-square',
-    },
-    'DISCORD CHATS': 'message-square',
-    'DISCORD RETURNS': 'message-square',
-    'DOWNLOADS': 'download',
-    'DRAFT NATIVE MESSAGES': 'message-circle',
-    'DUCKDUCKGO': {
-        'DUCKDUCKGO TAB THUMBNAILS': 'image',
-        'default': 'layers'
-    },
-    'EMULATED STORAGE METADATA': 'database',
-    'ENCRYPTING MEDIA APPS': 'lock',
-    'ETC HOSTS': 'globe',
-    'FACEBOOK MESSENGER': {
-        'CALLS': 'phone-call',
-        'CHAT': 'message-circle',
-        'CONTACTS': 'users',
-        'default': 'facebook',
-        '_mode': 'search',
-    },
-    'FACEBOOK - INSTAGRAM RETURNS': 'facebook',
-    'FILES APP': 'file-text',
-    'FILES BY GOOGLE': 'file',
-    'FIREBASE CLOUD MESSAGING': 'database',
-    'FIREFOX': {
-        'BOOKMARKS': 'bookmark',
-        'COOKIES': 'info',
-        'DOWNLOADS': 'download',
-        'FORM HISTORY': 'edit-3',
-        'PERMISSIONS': 'sliders',
-        'RECENTLY CLOSED TABS': 'x-square',
-        'SEARCH TERMS': 'search',
-        'TOP SITES': 'list',
-        'VISITS': 'globe',
-        'WEB HISTORY': 'globe',
-        'default': 'firefox',
-        '_mode': 'search',
-    },
-    'FITBIT': 'watch',
-    'GALLERY TRASH': 'image',
-    'GARMIN': {
-        'DEVICES': 'watch',
-        'NOTIFICATIONS': 'bell',
-        'SLEEP': 'moon',
-        'WEATHER': 'sun',
-        'default': 'activity',
-        '_mode': 'search',
-    },
-    'GARMIN-API': {
-        'ACTIVITY API': 'watch',
-        'DAILIES API': 'calendar',
-        'HEART RATE API': 'heart',
-        'STEPS API': 'arrow-up-circle',
-        'SLEEP API': 'moon',
-        'STRESS API': 'frown',
-        'POLYLINE API': 'map-pin',
-        'default': 'activity',
-        '_mode': 'search',
-    },
-    'GARMIN-CACHE': {
-        'ACTIVITIES': 'watch',
-        'CHARTS': 'activity',
-        'DAILIES': 'calendar',
-        'POLYLINE': 'map-pin',
-        'RESPONSE': 'terminal',
-        'SPO2': 'heart',
-        'SLEEP': 'moon',
-        'WEIGHT': 'bar-chart-2',
-        'default': 'activity',
-        '_mode': 'search',
-    },
-    'GARMIN-FILES': {
-        'LOG': 'file-text',
-        'PERSISTENT': 'code',
-        '_mode': 'search',
-    },
-    'GARMIN-GCM': {
-        'ACTIVITIES': 'watch',
-        'JSON': 'code',
-        '_mode': 'search',
-    },
-    'GARMIN-NOTIFICATIONS': 'message-square',
-    'GARMIN-SHAREDPREFS': {
-        'FACEBOOK': 'facebook',
-        'USER': 'user',
-        '_mode': 'search',
-    },
-    'GARMIN-SYNC': 'loader',
-    'GEO LOCATION': 'map-pin',
-    'GEOLOCATION': {
-        'APPLICATIONS': 'grid',
-        'MAP TILE CACHE': 'map',
-        'MAPSSYNC': 'map',
-        'PD PLACE CACHE': 'map-pin',
-        'default': 'map-pin',
-    },
-    'GMAIL': {
-        'GMAIL - LABEL DETAILS': 'mail',
-        'GMAIL - OFFLINE SEARCH': 'search',
-        'ACTIVE': 'at-sign',
-        'APP EMAILS': 'at-sign',
-        'DOWNLOAD REQUESTS': 'download-cloud',
-        'LABEL DETAILS': 'mail',
-        '_mode': 'search',
-    },
-    'GOOGLE CALL SCREEN': 'phone-incoming',
-    'GOOGLE CHAT': {
-        'GROUP INFORMATION': 'users',
-        'MESSAGES': 'message-circle',
-        'DRAFTS': 'edit-3',
-        'USERS': 'users',
-        'default': 'message-circle',
-        '_mode': 'search',
-    },
-    'GOOGLE DRIVE': 'file',
-    'GOOGLE DUO': {
-        'GOOGLE DUO - CALL HISTORY': 'phone-call',
-        'GOOGLE DUO - CLIPS': 'video',
-        'GOOGLE DUO - CONTACTS': 'user',
-        'CALL HISTORY': 'phone-call',
-        'CONTACTS': 'users',
-        'NOTES': 'edit-3',
-        '_mode': 'search',
-    },
-    'GOOGLE FIT (GMS)': 'activity',
-    'GOOGLE KEEP': 'list',
-    'GBOARD KEYBOARD': {
-        'CLIPBOARD': 'clipboard',
-        '_mode': 'search',
-    },
-    'GOOGLE MAPS VOICE GUIDANCE': 'map',
-    'GOOGLE MAPS TEMP VOICE GUIDANCE': 'map',
-    'GOOGLE MESSAGES': 'message-circle',
-    'GOOGLE NOW & QUICKSEARCH': 'search',
-    'GOOGLE PHOTOS': {
-        'LOCAL TRASH': 'trash-2',
-        'BACKED UP FOLDER': 'refresh-cw',
-        '_mode': 'search',
-    },
-    'GOOGLE PLAY': {
-        'GOOGLE PLAY SEARCHES': 'search',
-        'default': 'play',
-        '_mode': 'search',
-    },
-    'GOOGLE RETURNS': {
-        'GOOGLE RETURNS - ACTIVITIES': 'activity',
-        'default': 'chrome'
-    },
-    'GOOGLE RETURNS MBOXES': 'mail',
-    'GOOGLE RETURNS SUBSCRIBER INFO': 'user',
-    'GOOGLE RETURNS PLAY USER ACT': 'smartphone',
-    'GOOGLE RETURNS ANDROID DEVICE CONFIG': 'smartphone',
-    'GOOGLE RETURNS MY ACTIVITY IMAGE SEARCH': 'search',
-    'GOOGLE RETURNS MY ACTIVITY SEARCH': 'search',
-    'GOOGLE RETURNS YOUTUBE SUBS INFO': 'search',
-    'GOOGLE RETURNS ACCOUNT TARGET ASSOC. PHONE': 'target',
-    'GOOGLE RETURNS ACCOUNT TARGET ASSOC. COOKIES': 'target',
-    'GOOGLE RETURNS GOOGLE PLAY STORE DEVICES': 'tablet',
-    'GOOGLE RETURNS GOOGLE PLAY INSTALLS': 'shield',
-    'GOOGLE RETURNS GOOGLE PLAY LIBRARY': 'book-open',
-    'GOOGLE RETURNS GOOGLE PLAY USER REVIEWS': 'book-open',
-    'GOOGLE TAKEOUT ARCHIVE': {
-        'CHROME WEB HISTORY': 'chrome',
-        'CHROME ARC PACKAGES': 'package',
-        'CHROME AUTOFILL': 'edit-3',
-        'CHROME BOOKMARKS': 'star',
-        'CHROME DEVICE INFO': 'chrome',
-        'CHROME EXTENSIONS': 'tool',
-        'CHROME OS SETTINGS': 'settings',
-        'CHROME READING LIST': 'book',
-        'CHROME SEARCH ENGINES': 'search',
-        'CHROME OMNIBOX': 'search',
-        'FITBIT ACCOUNT PROFILE': 'user',
-        'FITBIT ACTIVITY GOALS': 'check-circle',
-        'FITBIT COMPUTED TEMPERATURE': 'thermometer',
-        'FITBIT OXYGEN SATURATION': 'droplet',
-        'FITBIT SLEEP': 'moon',
-        'FITBIT STRESS': 'activity',
-        'FITBIT TRACKERS': 'watch',
-        'GOOGLE ACCESS LOG ACTIVITIES': 'activity',
-        'GOOGLE ACCESS LOG DEVICES': 'smartphone',
-        'GOOGLE CHAT - MESSAGES': 'message-square',
-        'GOOGLE FI - USER INFO RECORDS': 'phone',
-        'GOOGLE FIT - DAILY ACTIVITY METRICS': 'trending-up',
-        'GOOGLE LOCATION HISTORY - LOCATION HISTORY': 'map-pin',
-        'GOOGLE PAY TRANSACTIONS': 'credit-card',
-        'GOOGLE PLAY STORE DEVICES': 'smartphone',
-        'GOOGLE PLAY STORE INSTALLS': 'box',
-        'GOOGLE PLAY STORE LIBRARY': 'grid',
-        'GOOGLE PLAY STORE PURCHASE HISTORY': 'shopping-cart',
-        'GOOGLE PLAY STORE REVIEWS': 'edit-3',
-        'GOOGLE PLAY STORE SUBSCRIPTIONS': 'refresh-cw',
-        'GOOGLE PROFILE': 'user',
-        'GOOGLE SEMANTIC LOCATION HISTORY - PLACE VISITS': 'map-pin',
-        'GOOGLE SEMANTIC LOCATION HISTORY - ACTIVITY SEGMENTS': 'activity',
-        'GOOGLE TASKS': 'check-circle',
-        'MBOX': 'mail',
-        'SAVED LINKS - DEFAULT LIST': 'list',
-        'SAVED LINKS - FAVORITE IMAGES': 'image',
-        'SAVED LINKS - FAVORITE PAGES': 'link-2',
-        'SAVED LINKS - WANT TO GO': 'navigation-2',
-        'YOUTUBE SUBSCRIPTIONS': 'youtube',
-        'default': 'user'
-    },
-    'GOOGLE TAKEOUT SEMANTIC LOCATIONS BY MONTH': 'map-pin',
-    'GOOGLE TASKS': 'list',
-    'GPS_DATA': 'map-pin',
-    'GROUPME': {
-        'GROUP INFORMATION': 'users',
-        'CHAT INFORMATION': 'message-circle',
-        '_mode': 'search',
-    },
-    'HEALTH': {
-        'DEFAULT': 'heart',
-        'HEALTH - ACHIEVEMENTS': 'star',
-        'HEALTH - HEADPHONE AUDIO LEVELS': 'headphones',
-        'HEALTH - HEART RATE': 'activity',
-        'HEALTH - RESTING HEART RATE': 'activity',
-        'HEALTH - STEPS': 'activity',
-        'HEALTH - WORKOUTS': 'activity',
-    },
-    'HIDEX': 'eye-off',
-    'HIKVISION': {
-        'CCTV ACTIVITY': 'activity',
-        'CCTV CHANNELS': 'film',
-        'CCTV INFO': 'settings',
-        'USER CREATED MEDIA': 'video',
-        '_mode': 'search',
-    },
-    'ICLOUD DOCUMENTS FOLDERS': 'book-open',
-    'ICLOUD QUICK LOOK': 'file',
-    'ICLOUD RETURNS': {
-        'ICLOUD - ACCOUNT FEATURES': 'user',
-        'default': 'cloud'
-    },
-    'ICLOUD SHARED ALBUMS': 'cloud',
-    'IDENTIFIERS': 'file',
-    'IMAGE MANAGER CACHE': 'image',
-    'IMO': {
-        'IMO - ACCOUNT ID': 'user',
-        'IMO - MESSAGES': 'message-square',
-    },
-    'IMO HD CHAT': {
-        'IMO HD CHAT - CONTACTS': 'user',
-        'IMO HD CHAT - MESSAGES': 'message-circle',
-    },
-    'INSTAGRAM': {
-        'INSTAGRAM THREADS': 'message-square',
-        'INSTAGRAM THREADS CALLS': 'phone',
-    },
-    'INSTAGRAM ARCHIVE': {
-        'INSTAGRAM ARCHIVE - ACCOUNT INFO': 'user',
-        'INSTAGRAM ARCHIVE - PERSONAL INFO': 'user',
-        'default': 'instagram'
-    },
-    'INSTALLED APPS': 'package',
-    'INTENTS': 'command',
-    'INTERACTIONC': {
-        'ATTACHMENTS': 'paperclip',
-        'CONTACTS': 'user',
-    },
-    'K-9 MAIL': {
-        'K-9 MAIL - ACCOUNTS': 'at-sign',
-        'default': 'mail'
-    },
-    'KEYBOARD': {
-        'KEYBOARD APPLICATION USAGE': 'type',
-        'KEYBOARD DYNAMIC LEXICON': 'type',
-    },
-    'KIK': {
-        'KIK GROUP ADMINISTRATORS': 'user-plus',
-        'KIK LOCAL ACCOUNT': 'user-check',
-        'KIK MEDIA METADATA': 'file-plus',
-        'KIK MESSAGES': 'message-square',
-        'KIK PENDING UPLOADS': 'upload',
-        'KIK USERS': 'user',
-        'KIK USERS IN GROUPS': 'user',
-    },
-    'KIK RETURNS': {
-        'KIK - PROFILE PIC': 'image',
-        'default': 'file-text'
-    },
-    'KLEINANZEIGEN.DE APP': {
-        'KLEINANZEIGEN.DE - ACCOUNT DETAILS': 'user',
-        'KLEINANZEIGEN.DE - NON RESETTABLE RECENT SEARCHES CACHE': 'search',
-        'KLEINANZEIGEN.DE - RECENT SEARCHES CACHE': 'search',
-        'KLEINANZEIGEN.DE - MESSAGEBOX': 'message-square',
-        'default': 'search',
-        '_mode': 'search',
-    },
-    'KNOWLEDGEC': {
-        'KNOWLEDGEC BATTERY LEVEL': 'battery',
-        'KNOWLEDGEC DEVICE LOCKED': 'lock',
-        'KNOWLEDGEC PLUGGED IN': 'battery-charging',
-        'default': 'activity',
-    },
-    'LEAPP_REPORT': {
-        'default': 'git-commit',
-        '_mode': 'search',
-    },
-    'LIBRE TORRENT': 'download',
-    'LIFE360': {
-        'LIFE360 - CHAT MESSAGES': 'message-circle',
-        'LIFE360 - DEVICE BATTERY': 'battery',
-        'LIFE360 - LOCATIONS': 'map-pin',
-        'LIFE360 - PLACES': 'map-pin',
-    },
-    'LINE': {
-        'LINE - CONTACTS': 'user',
-        'LINE - MESSAGES': 'message-square',
-        'LINE - CALL LOGS': 'phone',
-    },
-    'LOCATION SERVICES CONFIGURATIONS': 'settings',
-    'LOCATIONS': {
-        'APPLE MAPS SEARCH HISTORY': 'search',
-        'default': 'map-pin',
-    },
-    'MAP-MY-WALK': {
-        'ACTIVITIES': 'map',
-        'USER': 'user',
-        '_mode': 'search',
-    },
-    'MASTODON': {
-        'ACCOUNT DETAILS': 'user',
-        'ACCOUNT SEARCHES': 'users',
-        'HASHTAG SEARCHES': 'hash',
-        'INSTANCE DETAILS': 'info',
-        'NOTIFICATIONS': 'bell',
-        'TIMELINE': 'activity',
-        '_mode': 'search',
-    },
-    'MEDIA LIBRARY': 'play-circle',
-    'MEDIA METADATA': 'file-plus',
-    'MEDICAL ID': 'thermometer',
-    'MEGA': 'message-circle',
-    'METAMASK': {
-        'BROWSER': 'globe',
-        'CONTACTS': 'users',
-        '_mode': 'search',
-    },
-    'MEWE': 'message-circle',
-    'MICROSOFT RETURNS': 'target',
-    'MICROSOFT TEAMS': {
-        'TEAMS CALL LOGS': 'phone',
-        'TEAMS CONTACT': 'users',
-        'TEAMS MESSAGES': 'message-square',
-        'TEAMS SHARED LOCATIONS': 'map-pin',
-        'TEAMS USER': 'user',
-    },
-    'MICROSOFT TEAMS - LOGS': {
-        'TEAMS LOCATIONS': 'map-pin',
-        'TEAMS MOTION': 'move',
-        'TEAMS POWER LOG': 'battery-charging',
-        'TEAMS STATE CHANGE': 'truck',
-        'TEAMS TIMEZONE': 'clock',
-    },
-    'MOBILE ACTIVATION LOGS': 'clipboard',
-    'MOBILE BACKUP': 'save',
-    'MOBILE CONTAINER MANAGER': 'save',
-    'MOBILE INSTALLATION LOGS': 'clipboard',
-    'MOBILE SOFTWARE UPDATE': 'refresh-cw',
-    'MS AUTHENTICATOR': 'users',
-    'MY FILES': {
-        'MY FILES DB - CACHE MEDIA': 'image',
-        '_mode': 'search',
-    },
-    'NETFLIX ARCHIVE': {
-        'NETFLIX - BILLING HISTORY': 'credit-card',
-        'NETFLIX - PROFILES': 'users',
-        'NETFLIX - IP ADDRESS LOGIN': 'log-in',
-        'NETFLIX - ACCOUNT DETAILS': 'users',
-        'NETFLIX - MESSAGES SENT BY NETFLIX': 'mail',
-        'NETFLIX - SEARCH HISTORY': 'search',
-        'default': 'tv'
-    },
-    'NETWORK USAGE': {
-        'APP_DATA': 'activity',
-        'CONNECTIONS': 'bar-chart',
-        'default': 'send',
-        '_mode': 'search',
-    },
-    'NIKE-RUN': {
-        'ACTIVITIES': 'watch',
-        'ACTIVITY ROUTE': 'map-pin',
-        'ACTIVITY MOMENTS': 'list',
-        'NOTIFICATIONS': 'bell',
-        '_mode': 'search',
-    },
-    'NOTES': 'file-text',
-    'NOTIFICATIONS': 'bell',
-    'NOW PLAYING': 'music',
-    'OFFLINE PAGES': 'cloud-off',
-    'PACKAGE PREDICTIONS': 'package',
-    'PAS_DEBUG': {
-        'SEND GPS CAN DATA': 'map-pin',
-        'DEV LOC RESULTS': 'map-pin',
-        'ROAD SPEED LIMITS': 'target',
-        'ACCESS POINT LIST': 'wifi',
-        'VEHICLE SPEED': 'trending-up',
-        'TRANSMISSION STATUS': 'corner-up-right',
-        'OUTSIDE TEMPERATURE': 'thermometer',
-        'ODOMETER': 'plus-circle',
-        'default': 'archive',
-    },
-    'PERMISSIONS': 'check',
-    'PHONE BOOK DB': 'smartphone',
-    'PHONE CONFIG': 'smartphone',
-    'PHOTOS': {
-        'MIGRATIONS': 'chevrons-up',
-        'default': 'image',
-    },
-    'PIKPAK': 'cloud',
-    'PINGER': {
-        'PINGER - CDR': 'phone',
-        'PINGER - DML': 'phone',
-        'PINGER - IP': 'monitor',
-        'PINGER - MESSAGES': 'message-square',
-        'PINGER - ACCOUNT': 'user',
-        'default': 'phone'
-    },
-    'PLAYGROUND VAULT': 'lock',
-    'PODCAST ADDICT': 'music',
-    'POWER EVENTS': {
-        'POWER OFF RESET': 'power',
-        'LAST BOOT TIME': 'power',
-        'SHUTDOWN CHECKPOINTS': 'power',
-        '_mode': 'search',
-    },
-    'POWERLOG': 'power',
-    'POWERLOG BACKUPS': 'power',
-    'PREFERENCES PLIST': 'file',
-    'PRIVACY DASHBOARD': 'eye',
-    'PROTON MAIL': 'mail',
-    'PROTONMAIL': {
-        'CONTACTS': 'users',
-        'MESSAGES': 'inbox',
-        '_mode': 'search',
-    },
-    'PROTONVPN': 'shield',
-    'PUMA-TRAC': {
-        'ACTIVITIES': 'watch',
-        'USER': 'user',
-        '_mode': 'search',
-    },
-    'RAR LAB PREFS': 'settings',
-    'RCS CHATS': 'message-circle',
-    'RECENT ACTIVITY': 'activity',
-    'REDDIT RETURNS': 'chevrons-up',
-    'REMINDERS': 'list',
-    'ROUTINED': 'map',
-    'RUNKEEPER': {
-        'ACTIVITIES': 'watch',
-        'USER': 'user',
-        '_mode': 'search',
-    },
-    'SAFARI BROWSER': 'compass',
-    'SAMSUNG SMARTTHINGS': 'bluetooth',
-    'SAMSUNG WEATHER CLOCK': {
-        'DAILY': 'sunrise',
-        'HOURLY': 'thermometer',
-        '_mode': 'search',
-    },
-    'SAMSUNG BROWSER': 'globe',
-    'SAMSUNG_CMH': 'disc',
-    'SCREENTIME': 'monitor',
-    'SCRIPT LOGS': 'archive',
-    'SECRET CALCULATOR PHOTO ALBUM': 'image',
-    'SETTINGS SERVICES': 'battery-charging',
-    'SETTINGS SERVICES - BATTERY USAGE V9 - BATTERY STATES': 'battery-charging',
-    'SETTINGS SERVICES - BATTERY USAGE V9 - APP BATTERY USAGE EVENTS': 'bar-chart-2',
-    'SIM INFO': 'info',
-    'SKOUT': {
-        'SKOUT MESSAGES': 'message-circle',
-        'SKOUT USERS': 'users',
-    },
-    'SKYPE': {
-        'SKYPE - CALL LOGS': 'phone',
-        'SKYPE - MESSAGES': 'message-square',
-        'SKYPE - CONTACTS': 'user',
-    },
-    'SLACK': {
-        'SLACK ATTACHMENTS': 'paperclip',
-        'SLACK CHANNEL DATA': 'slack',
-        'SLACK MESSAGES': 'message-square',
-        'SLACK TEAM DATA': 'slack',
-        'SLACK USER DATA': 'user',
-        'SLACK WORKSPACE DATA': 'slack',
-    },
-    'SLOPES': {
-        'SLOPES - ACTIONS': 'trending-down',
-        'SLOPES - LIFT DETAILS': 'shuffle',
-        'SLOPES - RESORT DETAILS': 'home',
-    },
-    'SMS & IMESSAGE': 'message-square',
-    'SMS & MMS': 'message-square',
-    'SMS & MMS FROM BACKUP.AB': 'message-square',
-    'SNAPCHAT': 'bell',
-    'SNAPCHAT ARCHIVE': 'camera',
-    'SNAPCHAT RETURNS': 'camera',
-    'SQLITE JOURNALING': 'book-open',
-    'STRAVA': 'map',
-    'TEAMS': {  # TODO: align I & A artifacts since theres a 'microsoft teams' also
-        'TEAMS MESSAGES': 'message-circle',
-        'TEAMS USERS': 'users',
-        'TEAMS CALL LOG': 'phone',
-        'TEAMS ACTIVITY FEED': 'at-sign',
-        'TEAMS FILE INFO': 'file',
-        'default': 'file-text',
-    },
-    'TANGO': 'message-square',
-    'TELEGRAM': 'message-square',
-    'TELEGUARD': {
-        'TELEGUARD CONTACTS': 'users',
-        'TELEGUARD MESSAGES': 'message-square',
-        'TELEGUARD POSTS': 'at-sign',
-        'TELEGUARD CHANNELS': 'home',
-    },
-    'TELEMATICS': {
-        'GPS DETAIL': 'map-pin',
-        'WDWSTATUS REPORT': 'map-pin',
-        '_mode': 'search',
-    },
-    'TEXT INPUT MESSAGES': 'message-square',
-    'TEXT NOW': {
-        'TEXT NOW - CALL LOGS': 'phone',
-        'TEXT NOW - MESSAGES': 'message-square',
-        'TEXT NOW - CONTACTS': 'user',
-    },
-    'TIKTOK': {  # TODO: align I & A artifacts with each other
-        'TIKTOK CONTACTS': 'user',
-        'TIKTOK MESSAGES': 'message-square',
-        'TIKTOK SEARCH': 'search',
-        'TIKTOK - MESSAGES': 'message-square',
-        'TIKTOK - CONTACTS': 'user',
-    },
-    'TIKTOK RETURNS': 'film',
-    'TODOIST': {
-        'ITEMS': 'list',
-        'NOTES': 'file-text',
-        'PROJECTS': 'folder',
-        '_mode': 'search',
-    },
-    'TOR': 'globe',
-    'TORRENT DATA': 'download',
-    'TUSKY': {
-        'TIMELINE': 'activity',
-        'ACCOUNT': 'user',
-        '_mode': 'search',
-    },
-    'TWITTER': 'twitter',
-    'TWITTER RETURNS': 'twitter',
-    'USAGE STATS': 'bar-chart-2',
-    'USER DICTIONARY': 'book',
-    'VEHICLE INFO': 'truck',
-    'VENMO': 'dollar-sign',
-    'VERIZON RDD ANALYTICS': {
-        'VERIZON RDD - BATTERY HISTORY': 'power',
-        'VERIZON RDD - WIFI DATA': 'wifi',
-    },
-    'VIBER': {
-        'VIBER - CALL REMNANTS': 'phone-call',
-        'VIBER - CHATS': 'message-square',
-        'VIBER - CONTACTS': 'users',
-        'VIBER - SETTINGS': 'settings',
-        'VIBER - MESSAGES': 'message-square',
-        'VIBER - CALL LOGS': 'phone',
-    },
-    'VIPPS': {
-        'VIPPS CONTACTS': 'users',
-        'default': 'dollar-sign',
-    },
-    'VLC': {
-        'VLC MEDIA LIST': 'film',
-        'VLC THUMBNAILS': 'image',
-    },
-    'VLC THUMBS': {
-        'VLC MEDIA LIB': 'film',
-        'VLC THUMBNAILS': 'image',
-        'VLC THUMBNAIL DATA': 'image',
-        'default': 'image',
-    },
-    'VOICE-RECORDINGS': 'mic',
-    'VOICE-TRIGGERS': 'mic',
-    'WAZE': 'navigation-2',
-    'WHATSAPP': {  # TODO: I don't think search mode is required
-        'WHATSAPP - CONTACTS': 'users',
-        'WHATSAPP - MESSAGES': 'message-square',
-        'WHATSAPP - ONE TO ONE MESSAGES': 'message-circle',
-        'WHATSAPP - GROUP MESSAGES': 'message-circle',
-        'WHATSAPP - CALL LOGS': 'phone',
-        'WHATSAPP - GROUP DETAILS': 'users',
-        'default': 'user',
-        '_mode': 'search',
-    }, 
-    'WITHINGS HEALTH MATE' : {
-        'WITHINGS HEALTH MATE - ACCOUNTS': 'user',
-        'WITHINGS HEALTH MATE - CONTACTS': 'users',
-        'WITHINGS HEALTH MATE - DEVICES': 'watch',
-        'WITHINGS HEALTH MATE - LOCATIONS': 'map-pin',
-        'WITHINGS HEALTH MATE - TRACKINGS': 'activity',
-        'WITHINGS HEALTH MATE - MEASUREMENTS': 'activity',
-    },
-    'WHATSAPP EXPORTED CHAT': 'message-circle',
-    'WIFI CONNECTIONS': 'wifi',
-    'WIFI KNOWN NETWORKS': 'wifi',
-    'WIFI PROFILES': 'wifi',
-    'WIPE & SETUP': {
-        'FACTORY RESET': 'loader',
-        'SUGGESTIONS.XML': 'loader',
-        'SETUP_WIZARD_INFO.XML': 'loader',
-        'APPOPS.XML': 'loader',
-        'SAMSUNG WIPE HISTORY': 'trash-2',
-        'SAMSUNG WIPE RECOVERY HISTORY LOG': 'trash-2',
-        'default': 'loader',
-    },
-}
-
-# function that can be run against the list to sort and output to console
-def sort_and_print_mappings():
-    sorted_keys = sorted(icon_mappings.keys(), key=lambda x: x.lower())
-    sorted_dict = {key: icon_mappings[key] for key in sorted_keys}
-
-    print("{")
-    for category, mappings in sorted_dict.items():
-        if isinstance(mappings, dict):
-            print(f"    '{category}': {{")
-            # Sort the artifacts, with 'default' and '_mode' at the end
-            sorted_artifacts = sorted(
-                [(k, v) for k, v in mappings.items() if k not in ['default', '_mode']],
-                key=lambda x: x[0]
-            )
-            # Append 'default' and '_mode' at the end if they exist
-            if 'default' in mappings:
-                sorted_artifacts.append(('default', mappings['default']))
-            if '_mode' in mappings:
-                sorted_artifacts.append(('_mode', mappings['_mode']))
-            for artifact, icon in sorted_artifacts:
-                print(f"        '{artifact}': '{icon}',")
-            print("    },")
-        else:
-            print(f"    '{category}': '{mappings}',")
-    print("}")
-
-
-if __name__ == '__main__':
-    # Call the function to print the sorted mappings to the console
-    sort_and_print_mappings()
+from scripts.report_icons import icon_mappings, feather_icon_names
 
 def get_icon_name(category, artifact):
     """
@@ -962,7 +64,7 @@ def get_search_mode_categories():
 search_set = get_search_mode_categories()
 
 
-def generate_report(reportfolderbase, time_in_secs, time_HMS, extraction_type, image_input_path, casedata):
+def generate_report(reportfolderbase, time_in_secs, time_HMS, extraction_type, image_input_path, casedata, profile_filename, icons):
     control = None
     side_heading = \
         """
@@ -987,40 +89,40 @@ def generate_report(reportfolderbase, time_in_secs, time_HMS, extraction_type, i
     for root, dirs, files in sorted(os.walk(reportfolderbase)):
         files = sorted(files)
         for file in files:
+            if file.startswith('._'):
+                continue
             if file.endswith(".temphtml"):
                 fullpath = (os.path.join(root, file))
                 head, tail = os.path.split(fullpath)
-                p = pathlib.Path(fullpath)
+                filename = tail.replace(".temphtml", "")
+                p = Path(fullpath)
                 SectionHeader = (p.parts[-2])
                 if SectionHeader == '_elements':
                     pass
                 else:
-                    if control == SectionHeader:
-                        side_list[SectionHeader].append(fullpath)
-                        icon = get_icon_name(SectionHeader, tail.replace(".temphtml", ""))
-                        nav_list_data += list_item.format('', tail.replace(".temphtml", ".html"), icon,
-                                                          tail.replace(".temphtml", ""))
-                    else:
+                    if control != SectionHeader:
                         control = SectionHeader
                         side_list[SectionHeader] = []
-                        side_list[SectionHeader].append(fullpath)
                         nav_list_data += side_heading.format(SectionHeader)
-                        icon = get_icon_name(SectionHeader, tail.replace(".temphtml", ""))
-                        nav_list_data += list_item.format('', tail.replace(".temphtml", ".html"), icon,
-                                                          tail.replace(".temphtml", ""))
+                    side_list[SectionHeader].append(fullpath)
+                    icon_name = icons.get(SectionHeader, {}).get(filename, "")
+                    icon = icon_name if icon_name else get_icon_name(SectionHeader, filename)
+                    icon = icon if icon in feather_icon_names else 'alert-triangle'
+                    nav_list_data += list_item.format('', tail.replace(".temphtml", ".html").replace(" ", "_"), 
+                                                      icon, filename.replace("_", " "))
 
     # Now that we have all the file paths, start writing the files
 
-    for path_list in side_list.values():
+    for category, path_list in side_list.items():
         for path in path_list:
             old_filename = os.path.basename(path)
-            filename = old_filename.replace(".temphtml", ".html")
+            filename = old_filename.replace(".temphtml", ".html").replace(" ", "_")
             # search for it in nav_list_data, then mark that one as 'active' tab
             active_nav_list_data = mark_item_active(nav_list_data, filename) + nav_bar_script
             artifact_data = get_file_content(path)
 
             # Now write out entire html page for artifact
-            f = open(os.path.join(reportfolderbase, filename), 'w', encoding='utf8')
+            f = open(os.path.join(reportfolderbase, '_HTML', filename), 'w', encoding='utf8')
             artifact_data = insert_sidebar_code(artifact_data, active_nav_list_data, path)
             f.write(artifact_data)
             f.close()
@@ -1034,38 +136,22 @@ def generate_report(reportfolderbase, time_in_secs, time_HMS, extraction_type, i
                 pass # Perhaps it was not empty!
 
     # Create index.html's page content
-    create_index_html(reportfolderbase, time_in_secs, time_HMS, extraction_type, image_input_path, nav_list_data, casedata)
-    elements_folder = os.path.join(reportfolderbase, '_elements')
-    os.mkdir(elements_folder)
+    create_index_html(reportfolderbase, time_in_secs, time_HMS, extraction_type, image_input_path, nav_list_data, casedata, profile_filename)
+    elements_folder = os.path.join(reportfolderbase, '_HTML', '_elements')
     __location__ = os.path.dirname(os.path.abspath(__file__))
-    
-    shutil.copy2(os.path.join(__location__,"logo.png"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"ALEAPP_banner.png"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"dashboard.css"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"feather.min.js"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"dark-mode.css"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"dark-mode-switch.js"), elements_folder)
-    shutil.copytree(os.path.join(__location__,"MDB-Free_4.13.0"), os.path.join(elements_folder, 'MDB-Free_4.13.0'))
-    # Chart.js
-    shutil.copy2(os.path.join(__location__,"chart.umd.min.js"), elements_folder)
-    # Moment
-    shutil.copy2(os.path.join(__location__,"moment.min.js"), elements_folder)
-    # D3
-    shutil.copy2(os.path.join(__location__,"d3.v7.min.js"), elements_folder)
-    # Popper
-    shutil.copy2(os.path.join(__location__,"popper.min.js"), elements_folder)
-    # Cal-Heatmap
-    shutil.copy2(os.path.join(__location__,"cal-heatmap.css"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"cal-heatmap.min.js"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"Tooltip.min.js"), elements_folder)
-    # Highlight.js
-    shutil.copy2(os.path.join(__location__,"highlight.min.css"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"highlight.min.js"), elements_folder)
-    # Garmin Custom JS
-    shutil.copy2(os.path.join(__location__,"garmin-functions.js"), elements_folder)
-    shutil.copytree(os.path.join(__location__,"timeline"), os.path.join(elements_folder, 'timeline'))
-    shutil.copy2(os.path.join(__location__,"chat.css"), elements_folder)
-    shutil.copy2(os.path.join(__location__,"chat.js"), elements_folder)
+
+    def copy_no_perm(src, dst, *, follow_symlinks=True):
+        if not os.path.isdir(dst):
+            shutil.copyfile(src, dst)
+        return dst
+
+    try:
+        shutil.copytree(os.path.join(__location__, "_elements"), elements_folder, copy_function=copy_no_perm)
+    except shutil.Error:
+        print("shutil reported an error. Maybe due to recursive directory copying.")
+        if os.path.exists(os.path.join(elements_folder, 'MDB-Free_4.13.0')):
+            print("_elements folder seems fine. Probably nothing to worry about")
+
 
 def get_file_content(path):
     f = open(path, 'r', encoding='utf8')
@@ -1073,7 +159,7 @@ def get_file_content(path):
     f.close()
     return data
 
-def create_index_html(reportfolderbase, time_in_secs, time_HMS, extraction_type, image_input_path, nav_list_data, casedata):
+def create_index_html(reportfolderbase, time_in_secs, time_HMS, extraction_type, image_input_path, nav_list_data, casedata, profile_filename):
     '''Write out the index.html page to the report folder'''
     case_list = []
     content = '<br />'
@@ -1086,6 +172,9 @@ def create_index_html(reportfolderbase, time_in_secs, time_HMS, extraction_type,
         for key, value in casedata.items():
             if value:
                 case_list.append([key, value])
+    
+    if profile_filename:
+        case_list.append(['Profile loaded', profile_filename])
     
     case_list += [
         ['Extraction location', image_input_path],
@@ -1127,18 +216,35 @@ def create_index_html(reportfolderbase, time_in_secs, time_HMS, extraction_type,
     body_description = 'ALEAPP is an open source project that aims to parse every known Android artifact for the purpose of forensic analysis.'
     active_nav_list_data = mark_item_active(nav_list_data, filename) + nav_bar_script
 
+    html_reportfolderbase = Path(reportfolderbase).joinpath('_HTML')
+    html_reportfolderbase.mkdir(exist_ok=True)
+    with html_reportfolderbase.joinpath(filename).open('w', encoding='utf8') as f:
+        f.write(page_header.format(page_title))
+        f.write(body_start.format(f"ALEAPP {aleapp_version}"))
+        f.write(body_sidebar_setup + active_nav_list_data + body_sidebar_trailer)
+        f.write(body_main_header + body_main_data_title.format(body_heading, body_description))
+        f.write(content)
+        f.write(thank_you_note)
+        f.write(credits_code)
+        f.write(body_main_trailer + body_end + nav_bar_script_footer + page_footer)
+
+    # Create Index Redirection Page
+    redirection = \
+        """
+        <html>
+            <head>
+                <meta http-equiv="refresh" content="0; url=_HTML/index.html" />
+                <title>iLEAPP Report</title>
+            </head>
+        </html>
+        """
     f = open(os.path.join(reportfolderbase, filename), 'w', encoding='utf8')
-    f.write(page_header.format(page_title))
-    f.write(body_start.format(f"ALEAPP {aleapp_version}"))
-    f.write(body_sidebar_setup + active_nav_list_data + body_sidebar_trailer)
-    f.write(body_main_header + body_main_data_title.format(body_heading, body_description))
-    f.write(content)
-    f.write(thank_you_note)
-    f.write(credits_code)
-    f.write(body_main_trailer + body_end + nav_bar_script_footer + page_footer)
+    f.write(redirection)
     f.close()
 
-def generate_authors_table_code(aleapp_contributors):
+
+
+def generate_authors_table_code(ileapp_contributors):
     authors_data = ''
     for author_name, blog, tweet_handle, git in aleapp_contributors:
         author_data = ''
@@ -1208,5 +314,6 @@ def mark_item_active(data, itemname):
     else:
         ret = data[0: pos] + " active" + data[pos:]
         return ret
+
